Chain Types and Watches Updated

+++ 9.0/ArchivingServer/src/java/it/intecs/pisa/archivingserver/rest/ManagerServlet.java
@@ -334,11 +334,14 @@ public class ManagerServlet extends ArchivingServerServlet {
             File webinfDir = new File(appDir, "WEB-INF");
             ftpConfigDir=new File(webinfDir, "FTPServer");
             FTPService ftpService=this.getFtpService();
+            if(ftpService !=null)
             ftpService.deleteServer();
             ftpService = FTPService.getInstance(ftpConfigDir.getAbsolutePath());
             this.setFtpService(ftpService);
             if(ftpService==null)
                 Log.log("Cannot start FTP server");
+            else
+                Log.log("FTP server started");
         }    
        
     }
